requirements.txt
langchain-community==0.0.20
langchain-core==0.1.22
langsmith==0.0.87
langchain
python-dotenv==1.0.1
streamlit==1.37.1
boto3>=1.33.2
botocore>=1.33.2
langchain_aws
psycopg2-binary
faiss-cpu